deleting some still not used files till they will have a useful reason to be there moved tests to related folder some issues to be fix in this branch
