+++ src/tests/jesttest.test.js
@@ -1,4 +1,4 @@
-const jesttest = require('./jesttest');
+const jesttest = require('../js/jesttest');
 
 test('adds 1 + 2 to equal 3', () => {
     expect(jesttest(1, 2)).toBe(3);
